+++ core/initializer/test_suite_metadata_acquirer/test_suite_metadata_acquirer.go
@@ -43,6 +43,7 @@ func GetTestSuiteMetadata(
 	logrus.Info("Launching metadata-acquiring containers...")
 	testsuiteContainerId, apiContainerId, err := launcher.LaunchMetadataAcquiringContainers(
 		parentContext,
+		logrus.StandardLogger(),
 		dockerManager,
 		suiteExecutionVolName,
 		debuggerHostPortBinding)